+++ src/components/chat/elements/Sidebar.vue
@@ -2,13 +2,16 @@
   <div
     class="vdb-c-flex vdb-c-h-full vdb-c-flex-col vdb-c-gap-16 vdb-c-border-r vdb-c-bg-white vdb-c-p-16 vdb-c-pl-12 vdb-c-pr-20 vdb-c-text-black"
   >
-    <div class="vdb-c-pt-12 vdb-c-text-2xl vdb-c-font-bold">
+    <div class="vdb-c-text-2xl vdb-c-font-bold">
       <component :is="config.icon" />
     </div>
     <Button
       variant="primary"
-      @click="$emit('create-new-session')"
       class="vdb-c-px-10 vdb-c-py-12"
+      :class="{
+        'vdb-c-pointer-events-none vdb-c-opacity-20': status === 'inactive',
+      }"
+      @click="$emit('create-new-session')"
     >
       <div class="vdb-c-flex vdb-c-items-center vdb-c-gap-6">
         <ComposeIcon />
@@ -20,6 +23,9 @@
     </Button>
     <div
       class="vdb-c-flex vdb-c-flex-grow vdb-c-flex-col vdb-c-gap-16 vdb-c-overflow-hidden"
+      :class="{
+        'vdb-c-pointer-events-none vdb-c-opacity-20': status === 'inactive',
+      }"
     >
       <!-- Explore Agents -->
       <div
@@ -45,7 +51,7 @@
           />
         </button>
         <div
-          v-if="showExploreAgents"
+          v-if="status !== 'inactive' && showExploreAgents"
           class="vdb-c-overflow-y-auto vdb-c-rounded-lg vdb-c-px-8 vdb-c-py-4"
         >
           <template v-for="(agent, index) in agents" :key="index">
@@ -80,23 +86,33 @@
                 'vdb-c-h-16 vdb-c-w-16 vdb-c-transition-transform',
                 { 'vdb-c-rotate-180': showSessions },
               ]"
-              :stroke-width="1.5"
+              stroke-color="#464646"
+              :stroke-width="2"
             />
           </div>
         </button>
-        <div v-if="showSessions" class="vdb-c-mt-4 vdb-c-overflow-y-auto">
-          <div v-if="addDummySession">
-            <div
-              class="vdb-c-cursor-pointer vdb-c-truncate vdb-c-rounded-lg vdb-c-bg-roy vdb-c-p-8 vdb-c-text-sm vdb-c-font-medium vdb-c-text-vdb-darkishgrey"
-            >
-              (new session)
+        <div
+          v-if="status !== 'inactive' && showSessions"
+          class="vdb-c-mt-4 vdb-c-overflow-y-auto"
+        >
+          <transition name="fade" mode="out-in">
+            <div v-if="addDummySession">
+              <div
+                class="vdb-c-cursor-pointer vdb-c-truncate vdb-c-rounded-lg vdb-c-bg-roy vdb-c-p-8 vdb-c-text-sm vdb-c-font-medium vdb-c-text-vdb-darkishgrey"
+              >
+                (new session)
+              </div>
             </div>
-          </div>
-          <template v-for="session in sessions" :key="session.session_id">
+          </transition>
+          <transition-group name="fade" tag="div">
             <div
+              v-for="session in sessions"
+              :key="session.session_id"
               @click="$emit('session-click', session.session_id)"
+              @mouseenter="hoveredSession = session.session_id"
+              @mouseleave="hoveredSession = null"
               :class="[
-                'vdb-c-cursor-pointer vdb-c-truncate vdb-c-rounded-lg vdb-c-p-8 vdb-c-text-sm vdb-c-font-medium vdb-c-text-vdb-darkishgrey',
+                'vdb-c-flex vdb-c-cursor-pointer vdb-c-items-center vdb-c-justify-between vdb-c-truncate vdb-c-rounded-lg vdb-c-p-8 vdb-c-px-12 vdb-c-text-sm vdb-c-font-medium vdb-c-text-vdb-darkishgrey',
                 {
                   'vdb-c-bg-roy': session.session_id === selectedSession,
                   'hover:vdb-c-bg-gray-100':
@@ -104,22 +120,34 @@
                 },
               ]"
             >
-              {{
-                new Date(session.created_at * 1000)
-                  .toLocaleString("en-US", {
-                    year: "numeric",
-                    month: "2-digit",
-                    day: "2-digit",
-                    hour: "2-digit",
-                    minute: "2-digit",
-                    second: "2-digit",
-                    hour12: false,
-                  })
-                  .replace(/\//g, ".")
-                  .replace(",", " -")
-              }}
+              <span>
+                {{
+                  new Date(session.created_at * 1000)
+                    .toLocaleString("en-US", {
+                      year: "numeric",
+                      month: "2-digit",
+                      day: "2-digit",
+                      hour: "2-digit",
+                      minute: "2-digit",
+                      second: "2-digit",
+                      hour12: false,
+                    })
+                    .replace(/\//g, ".")
+                    .replace(",", " -")
+                }}
+              </span>
+              <span
+                @click.stop="$emit('delete-session', session.session_id)"
+                class="vdb-c-transition-all vdb-c-duration-300 hover:vdb-c-scale-110"
+              >
+                <DeleteIcon
+                  :fill="
+                    hoveredSession === session.session_id ? 'black' : '#CCCCCC'
+                  "
+                />
+              </span>
             </div>
-          </template>
+          </transition-group>
         </div>
       </div>
 
@@ -141,11 +169,15 @@
                 'vdb-c-h-16 vdb-c-w-16 vdb-c-transition-transform',
                 { 'vdb-c-rotate-180': showCollections },
               ]"
-              :stroke-width="1.5"
+              stroke-color="#464646"
+              :stroke-width="2"
             />
           </div>
         </button>
-        <div v-if="showCollections" class="vdb-c-mt-4 vdb-c-overflow-y-auto">
+        <div
+          v-if="status !== 'inactive' && showCollections"
+          class="vdb-c-mt-4 vdb-c-overflow-y-auto"
+        >
           <template v-for="collection in collections" :key="collection.id">
             <div
               @click="$emit('collection-click', collection.id)"
@@ -171,14 +203,20 @@
       <a
         v-for="(link, index) in config.links"
         :key="index"
-        class="vdb-c-mx-8 vdb-c-my-12 vdb-c-text-sm vdb-c-font-medium vdb-c-text-vdb-darkishgrey hover:vdb-c-text-vdb-darkishgrey hover:vdb-c-no-underline"
+        class="vdb-c-mx-8 vdb-c-my-12 vdb-c-pl-8 vdb-c-text-sm vdb-c-font-medium vdb-c-text-vdb-darkishgrey hover:vdb-c-text-black hover:vdb-c-no-underline"
         :href="link.href"
         :target="link.target || '_blank'"
         rel="noopener noreferrer"
       >
         {{ link.text }}
       </a>
-      <Button class="vdb-c-w-full" variant="tertiary">
+      <Button
+        class="vdb-c-w-full"
+        variant="tertiary"
+        :class="{
+          'vdb-c-pointer-events-none vdb-c-opacity-20': status === 'inactive',
+        }"
+      >
         <div
           class="vdb-c-flex vdb-c-w-full vdb-c-items-center vdb-c-justify-center vdb-c-gap-6"
         >
@@ -193,12 +231,12 @@
                 v-if="typeof config.primaryLink.icon === 'string'"
                 :src="config.primaryLink.icon"
                 alt="Primary Link Icon"
-                class="vdb-c-mr-8 vdb-c-h-16 vdb-c-w-16"
+                class="vdb-c-mr-8 vdb-c-h-20 vdb-c-w-20"
               />
               <component
                 v-else-if="typeof config.primaryLink.icon === 'object'"
                 :is="config.primaryLink.icon"
-                class="vdb-c-mr-8 vdb-c-h-16 vdb-c-w-16"
+                class="vdb-c-mr-8 vdb-c-h-20 vdb-c-w-20"
               />
             </template>
             {{ config.primaryLink.text }}
@@ -214,46 +252,10 @@ import { nextTick, ref, watch, computed } from "vue";
 
 import Button from "../../buttons/Button.vue";
 
-import SpielbergIcon from "../../icons/Spielberg2.vue";
+import DeleteIcon from "../../icons/Delete.vue";
 import ComposeIcon from "../../icons/Compose.vue";
 import ChevronDown from "../../icons/ChevronDown.vue";
 import MenuIcon from "../../icons/Menu.vue";
-
-const showExploreAgents = ref(false);
-const showSessions = ref(true);
-const showCollections = ref(false);
-const isExploreAgentsFocused = ref(false);
-const exploreAgentsTimeout = ref(null);
-const userClickedSessions = ref(false);
-const userClickedCollections = ref(false);
-
-const toggleExploreAgents = (value) => {
-  showExploreAgents.value =
-    value !== undefined ? value : !showExploreAgents.value;
-};
-
-const toggleSessions = (value) => {
-  userClickedSessions.value = true;
-  showSessions.value = value !== undefined ? value : !showSessions.value;
-};
-
-const toggleCollections = (value) => {
-  userClickedCollections.value = true;
-  showCollections.value = value !== undefined ? value : !showCollections.value;
-};
-
-const triggerExploreAgentsFocusAnimation = () => {
-  if (exploreAgentsTimeout.value) {
-    clearTimeout(exploreAgentsTimeout.value);
-  }
-  isExploreAgentsFocused.value = false;
-  nextTick(() => {
-    isExploreAgentsFocused.value = true;
-    exploreAgentsTimeout.value = setTimeout(() => {
-      isExploreAgentsFocused.value = false;
-    }, 1000);
-  });
-};
 
 const props = defineProps({
   sessions: {
@@ -267,6 +269,10 @@ const props = defineProps({
   agents: {
     type: Array,
     required: true,
+  },
+  status: {
+    type: String,
+    default: "active",
   },
   config: {
     type: Object,
@@ -297,6 +303,43 @@ const props = defineProps({
     default: false,
   },
 });
+
+const showExploreAgents = ref(false);
+const showSessions = ref(true);
+const showCollections = ref(false);
+const isExploreAgentsFocused = ref(false);
+const exploreAgentsTimeout = ref(null);
+const userClickedSessions = ref(false);
+const userClickedCollections = ref(false);
+const hoveredSession = ref(null);
+
+const toggleExploreAgents = (value) => {
+  showExploreAgents.value =
+    value !== undefined ? value : !showExploreAgents.value;
+};
+
+const toggleSessions = (value) => {
+  userClickedSessions.value = true;
+  showSessions.value = value !== undefined ? value : !showSessions.value;
+};
+
+const toggleCollections = (value) => {
+  userClickedCollections.value = true;
+  showCollections.value = value !== undefined ? value : !showCollections.value;
+};
+
+const triggerExploreAgentsFocusAnimation = () => {
+  if (exploreAgentsTimeout.value) {
+    clearTimeout(exploreAgentsTimeout.value);
+  }
+  isExploreAgentsFocused.value = false;
+  nextTick(() => {
+    isExploreAgentsFocused.value = true;
+    exploreAgentsTimeout.value = setTimeout(() => {
+      isExploreAgentsFocused.value = false;
+    }, 1000);
+  });
+};
 
 const computedSelectedCollection = computed(() => {
   if (props.selectedCollection !== "default") {
@@ -334,6 +377,7 @@ watch(showExploreAgents, (newValue) => {
 defineEmits([
   "create-new-session",
   "session-click",
+  "delete-session",
   "collection-click",
   "agent-click",
 ]);
@@ -373,5 +417,15 @@ defineExpose({
   100% {
     background-color: transparent;
   }
+}
+
+.fade-enter-active,
+.fade-leave-active {
+  transition: opacity 0.5s ease;
+}
+
+.fade-enter-from,
+.fade-leave-to {
+  opacity: 0;
 }
 </style>