+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+### [0.0.5]() - 2024-10-23
+
+### Added
+- Setup Screen
+- Delete Session
+- Video View in session History
+- Expose ChatInputRef & createNewSession()
+
+### Fixes
+- Full Screen for <ChatVideo/> message handler
+- Open Directors Log only when status is not success
+- Minor tweak in Default Screen's search suggestions
+- Fix Sidebar's Icon anaimation
+
+
 ### [0.0.4]() - 2024-09-18
 
 ### Added